Title: 20대의 평균나이, Time: 1018ms, Memory: 0MB, Status: Time Limit Exceed - CodeTree

+++ 240221/20대의 평균나이/average-age-of-20.cpp
@@ -5,16 +5,16 @@ using namespace std;
 
 int main() {
     int age;
-    int total = 0;
+    float total = 0.0;
     int cnt = 0;
 
     while(1){
         cin >> age;
         if(age > 29) break;
 
-        total += age;
+        total += (float)age;
         cnt++;
     }
-    printf("%.2f", (float)total / cnt);
+    printf("%.2f", total / cnt);
     return 0;
 }
